+++ packages/client/src/types.ts
@@ -5,6 +5,7 @@ import type {
   Images,
   ProjectSettings,
   Cron,
+  Services,
   ExperimentalServices,
   ExperimentalServiceGroups,
 } from '@vercel/build-utils';
@@ -189,6 +190,7 @@ export interface VercelConfig {
    * This file will be included in prebuilt deployments.
    */
   bulkRedirectsPath?: string | null;
+  services?: Services;
   /**
    * @experimental This feature is experimental and may change.
    */