+++ packages/fastify/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/fastify
 
+## 0.1.79
+
+### Patch Changes
+
+- Updated dependencies [22f77b9]
+  - @vercel/node@5.8.0
+
+## 0.1.78
+
+### Patch Changes
+
+- Updated dependencies [f0d7d32]
+  - @vercel/node@5.7.17
+
 ## 0.1.77
 
 ### Patch Changes